+++ pytket/docs/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 Minor new features:
 * New properties: :py:meth:``circuit.Op.dagger`` and :py:meth:``circuit.Op.transpose``.
 * New ``NPhasedX`` OpType.
+* New ``GlobalPhasedXPredicate`` and ``GlobalisePhasedX`` (transform and pass).
 
 0.16.0 (October 2021)
 ---------------------